rename web to httpserver in code
